Test and ensure middleware pipeline works as expected
Removes usage of `RouteCollector` in favor of direct injection of
`Route` instances in the created router instance. Updates test
expectations to work with the returned middleware pipe instead of
assuming a factory produces an application instance.

In order to test appropriately, we need a concrete router so we can get
the class and `new` a new instance, so fast route is added as a dev
dependency.