docs: updates CHANGELOG entries for #3
This patch updates the CHANGELOG entries for #3 to reference the pull
request, so users can understand the source of the change.

Additionally, it pushes the new version entry to the top of the file
(CHANGELOG.md files list in reverse chronological order), sets the
version to 2.0.0, and sets the release date.